sources/manager/requirements.txt
#-e ../common/libs/remoulade/[rabbitmq,redis,postgres]
requests
agraph-python==101.0.7
urllib3==1.26.18 
uvicorn
FastAPI
gevent
rdflib
threadsafe-async
python-jose[cryptography]

